requirements.txt
easyocr
pillow
numpy
unidecode
